Add flag to ignore RSVP status for own events
I haven't been able to pinpoint the core reason, but I've observed that for events created by me sometimes `myStatus` is `null` and others it's `CalendarApp.GuestStatus.OWNER`. The latter causes the checks again YES/MAYBE/null to fail.

It's possible that folks have use cases where they don't want to sync events that they created but haven't explicitly RSVPed to (e.g. maybe one created an event for one's partner's appointment that they themselves are not attending), so I added this as a flag to allow customization.